// crm_frontend_3/src/api/api2.ts
export interface Create_RAMDto {
    name_Сomponent: string ;
    price: number;
    memory_type: string ;
    form_factor: string ;
    memory_module_key: string ;
    volume: string ;
    clock_frequency: string;
    timing: string ;
    name_Company: string ;
}